Remove edit
