+++ addon/locales/ru/components/geometry-add-modes/rhumb.js
@@ -1,6 +1,5 @@
 export default {
   'menu-button-tooltip': 'Добавить по румбу',
-  'menu-button-tooltip-edit': 'Редактировать по румбу',
   'dialog-approve-button-caption': 'Сохранить',
   'dialog-deny-button-caption': 'Отмена',
   'type-object-field-label': 'Тип объекта',
