+++ addon/templates/components/geometry-add-modes/rhumb.hbs
@@ -1,6 +1,6 @@
 {{flexberry-button
   tooltip=menuButtonTooltip
-  click=(action "onOpenDialog" tabModel)
+  click=(action "onOpenDialog")
   iconClass="flexberry-map-command-icon location map-pin icon"
 }}
 {{!
@@ -68,7 +68,7 @@
             <td class="narrow">
               {{flexberry-button
                 caption=addButtonCaption
-                click=(action "onAddRaw")
+                click=(action "onAddRow")
               }}
             </td>
           </tr>
